fix: Complete i32 type conversion for SQLite INT4 compatibility

+++ server/src/t2/ai.rs
@@ -275,7 +275,7 @@ match_score от 0 до 100 - насколько товар соответств
 
     #[derive(Deserialize)]
     struct AIRecommendation {
-        product_id: i64,
+        product_id: i32,
         price_category: String,
         match_score: i32,
         match_reasons: Vec<String>,
@@ -374,7 +374,7 @@ pub async fn recommend_accessories(
 
     #[derive(Deserialize)]
     struct AIAccessory {
-        product_id: i64,
+        product_id: i32,
         reason: String,
         benefit: String,
     }
@@ -467,7 +467,7 @@ pub async fn recommend_tariffs(
 
     #[derive(Deserialize)]
     struct AITariff {
-        tariff_id: i64,
+        tariff_id: i32,
         recommendation: String,
     }
